pyproject.toml
[tool.poetry]
name = "talk-with-chatgpt"
version = "0.1.0"
description = ""
authors = ["Tomohiko Fujino"]
readme = "README.md"
packages = [{include = "talk_with_chatgpt"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.8"
speechrecognition = "^3.10.0"
python-dotenv = "^1.0.0"
gtts = "^2.3.2"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
